+++ source/MaterialXCore/Definition.cpp
@@ -26,6 +26,8 @@ const string TypeDef::CONTEXT_ATTRIBUTE = "context";
 const string Implementation::FILE_ATTRIBUTE = "file";
 const string Implementation::FUNCTION_ATTRIBUTE = "function";
 const string Implementation::LANGUAGE_ATTRIBUTE = "language";
+const string UnitDef::UNITTYPE_ATTRIBUTE = "unittype";
+const string UnitTypeDef::DEFAULT_ATTRIBUTE = "default";
 
 //
 // NodeDef methods
@@ -145,6 +147,19 @@ NodeDefPtr Implementation::getNodeDef() const
 ConstNodeDefPtr Implementation::getDeclaration(const string&) const
 {
     return getNodeDef();
+}
+
+vector<UnitDefPtr> UnitTypeDef::getUnitDefs() const
+{
+    vector<UnitDefPtr> unitDefs;
+    for (UnitDefPtr unitDef : getDocument()->getChildrenOfType<UnitDef>())
+    {
+        if (unitDef->getUnitType() == _name)
+        {
+            unitDefs.push_back(unitDef);
+        }
+    }
+    return unitDefs;
 }
 
 } // namespace MaterialX