+++ source/MaterialXCore/Element.cpp
@@ -33,6 +33,8 @@ const string ValueElement::UI_FOLDER_ATTRIBUTE = "uifolder";
 const string ValueElement::UI_MIN_ATTRIBUTE = "uimin";
 const string ValueElement::UI_MAX_ATTRIBUTE = "uimax";
 const string ValueElement::UI_ADVANCED_ATTRIBUTE = "uiadvanced";
+const string ValueElement::UNIT_ATTRIBUTE = "unit";
+const string ValueElement::UNITTYPE_ATTRIBUTE = "unittype";
 
 Element::CreatorMap Element::_creatorMap;
 
@@ -584,6 +586,26 @@ ValuePtr ValueElement::getDefaultValue() const
     return ValuePtr();
 }
 
+const string& ValueElement::getActiveUnit() const
+{
+    // Return the unit, if any, stored in our declaration.
+    ConstElementPtr parent = getParent();
+    ConstInterfaceElementPtr interface = parent ? parent->asA<InterfaceElement>() : nullptr;
+    if (interface)
+    {
+        ConstNodeDefPtr decl = interface->getDeclaration();
+        if (decl)
+        {
+            ValueElementPtr value = decl->getActiveValueElement(getName());
+            if (value)
+            {
+                return value->getUnit();
+            }
+        }
+    }
+    return EMPTY_STRING;
+}
+
 bool ValueElement::validate(string* message) const
 {
     bool res = true;
@@ -614,6 +636,40 @@ bool ValueElement::validate(string* message) const
                 }
             }
         }
+    }
+    UnitTypeDefPtr unitTypeDef;
+    if (hasUnitType())
+    {
+        const string& unittype = getUnitType();
+        if (!unittype.empty())
+        {
+            unitTypeDef = getDocument()->getUnitTypeDef(unittype);
+            validateRequire(unitTypeDef != nullptr, res, message, "Unit type definition does not exist in document");
+        }
+    }            
+    if (hasUnit())
+    {
+        bool foundUnit = false;
+        if (unitTypeDef)
+        {
+            const string& unit = getUnit();
+            if (unitTypeDef->getDefault() == unit)
+            {
+                foundUnit = true;
+            }
+            else
+            {
+                for (UnitDefPtr unitDef : unitTypeDef->getUnitDefs())
+                {
+                    if (unitDef->getUnit(unit))
+                    {
+                        foundUnit = true;
+                        break;
+                    }
+                }
+            }
+        }
+        validateRequire(foundUnit, res, message, "Unit definition does not exist in document");
     }
     return TypedElement::validate(message) && res;
 }
@@ -760,6 +816,9 @@ INSTANTIATE_CONCRETE_SUBCLASS(PropertySetAssign, "propertysetassign")
 INSTANTIATE_CONCRETE_SUBCLASS(ShaderRef, "shaderref")
 INSTANTIATE_CONCRETE_SUBCLASS(Token, "token")
 INSTANTIATE_CONCRETE_SUBCLASS(TypeDef, "typedef")
+INSTANTIATE_CONCRETE_SUBCLASS(Unit, "unit")
+INSTANTIATE_CONCRETE_SUBCLASS(UnitDef, "unitdef")
+INSTANTIATE_CONCRETE_SUBCLASS(UnitTypeDef, "unittypedef")
 INSTANTIATE_CONCRETE_SUBCLASS(Variant, "variant")
 INSTANTIATE_CONCRETE_SUBCLASS(VariantAssign, "variantassign")
 INSTANTIATE_CONCRETE_SUBCLASS(VariantSet, "variantset")