+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [1.36.5] - Development
 
+### Added
+- Added an initial TextureBaker class, supporting baking of graph outputs to textures.
+- Added initial support for units, including the MaterialX\:\:Unit, MaterialX\:\:UnitDef, and MaterialX\:\:UnitTypeDef classes.
+- Added support for unit conversion in shader code generation.
+
+### Changed
+- Updated subsurface interface and logic for Autodesk Standard Surface.
+
 ## [1.36.4] - 2019-09-26
 
 ### Added